+++ app/common/controller/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseController
+
+__all__ = ["BaseController"]
