+++ app/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import settings
+
+__all__ = ["settings"]
