+++ app/controllers/__init__.py
@@ -1,1 +1,3 @@
+from .query import QueryController
 
+__all__ = ["QueryController"]
