+++ app/repositories/__init__.py
@@ -1,0 +1,3 @@
+from .query import QueryRepository
+
+__all__ = ["QueryRepository"]
