+++ app/schemas/extras/__init__.py
@@ -1,0 +1,4 @@
+from .common import Response
+from .health import Health
+
+__all__ = ["Response", "Health"]
