+++ app/schemas/requests/__init__.py
@@ -1,0 +1,3 @@
+from .query import SearchBodyRequest
+
+__all__ = ["SearchBodyRequest"]
