+++ app/services/__init__.py
@@ -1,0 +1,3 @@
+from .query import QueryService
+
+__all__ = ["QueryService"]
